refactor: organizing some files and fixing some bugs
