+++ backend/.gitignore
@@ -1,1 +1,3 @@
-/uploads/
+/uploads/*.png
+/uploads/*.jpeg
+/uploads/*.jpg
