+++ backend/src/config/game.config.ts
@@ -1,0 +1,6 @@
+// Configurações do jogo
+export const MAX_PLAYERS = 4;
+export const TURN_TIME_LIMIT = 30000; // 30 segundos por turno
+export const CARDS_PER_PLAYER = 5;
+export const CARDS_PER_TYPE = 6; // 6 cartas de cada tipo (rei, dama, ás)
+export const JOKERS_COUNT = 2; // 2 coringas
