+++ backend/src/routes/lobby.routes.ts
@@ -1,10 +1,10 @@
-import express from 'express';
-import { CreateHandler, EnterHandler } from '../controllers/lobby.controllers';
+// import express from 'express';
+// import { CreateHandler, EnterHandler } from '../controllers/lobby.controllers';
 
-const router = express.Router();
+// const router = express.Router();
 
 //Defines the API endpoints to create or enter a game Room
-router.post('/create', CreateHandler);
-router.post('/enter', EnterHandler)
+// router.post('/create', CreateHandler);
+// router.post('/enter', EnterHandler)
 
-export default router
+// export default router
